pawgo/requirements.txt
asgiref==3.7.2
Django==4.2
sqlparse==0.4.4
tzdata==2023.3
django-money
django-phonenumber-field[phonenumbers]
